+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './constants';
+export * from './isNumericQuantity';
 export * from './numericQuantity';
 export * from './parseRomanNumerals';
 export * from './types';